backend/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
google-cloud-storage
reportlab
python-dotenv
alembic
uvicorn
